:arrow_up: Fix Google provider to be more compatible
Using the `MAJOR.MINOR.PATCH` method ~> `MAJOR.MINOR.PATCH` will only allow the PATCH value to increment. This becomes a problem when the terraform is using a new MINOR version.  When that happens the dependency conflict cannot be resolved.  This updates the Google Provider to the latest version, and also changes it to only use `MAJOR.MINOR`. This will allow terraform using this module to increment the MINOR version as needed.

Story: [sc-256414]