Add Project to resources and Provider
The some resources in this module are acctully from the google-beta provider, but it has not been set as a required provider. This sets it. In addition we add the project field to several resources because the lastest version of the google provider seems to require it. It is also possible that it is just required when using two instances of the module in the same terraform state.